// src/app/generated/caster-api/model/fileVersion.ts
// Copyright 2021 Carnegie Mellon University. All Rights Reserved.
// Released under a MIT (SEI)-style license. See LICENSE.md in the project root for license information.

/**
 * Caster API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FileVersion { 
    /**
     * ID of the file version.
     */
    id?: string;
    /**
     * ID of the file.
     */
    fileId?: string | null;
    /**
     * Name of the file.
     */
    name?: string | null;
    /**
     * The ID of the user who saved the file last.
     */
    modifiedById?: string | null;
    /**
     * The name of the user who saved the file last.
     */
    modifiedByName?: string | null;
    /**
     * The date the file was saved.
     */
    dateSaved?: Date | null;
    /**
     * The full contents of the file.
     */
    content?: string | null;
    /**
     * Tag string applied to this version of the file.
     */
    tag?: string | null;
    /**
     * The ID of the user who tagged the file.
     */
    taggedById?: string | null;
    /**
     * The name of the user who tagged the file.
     */
    taggedByName?: string | null;
    /**
     * Date the tag was applied to this version of the file.
     */
    dateTagged?: Date | null;
}